Extracted post messages.
